+++ Warrion.sol
@@ -7,7 +7,7 @@ import "BaseStation.sol";
 import "WarUnit.sol";
 
 contract Warrior is WarUnit{
-    constructor(BaseStation baseStationAddress, uint8 warriorHealth) public WarUnit(baseStationAddress, warriorHealth){
+    constructor(BaseStation baseStationAddress, uint16 warriorHealth) public WarUnit(baseStationAddress, warriorHealth){
         require(tvm.pubkey() != 0, 101);
         require(msg.pubkey() == tvm.pubkey(), 102);
         tvm.accept();
